Add a Aptly::Publishable module which gets included in Aptly::Repository and Aptly::Snapshot
