+++ lib/aptly/snapshot.rb
@@ -1,9 +1,11 @@
 require_relative 'representation'
-
+require_relative 'publishable'
 module Aptly
   # Aptly snapshots representation.
   # @see http://www.aptly.info/doc/api/snapshots/
   class Snapshot < Representation
+    include Publishable
+
     # Updates this snapshot
     # @return [self] if the instance data was mutated
     # @return [nil] if the instance data was not mutated
@@ -48,24 +50,6 @@ module Aptly
     # @return [PublishedRepository] newly published repository
     def publish(prefix, **kwords)
       Aptly.publish([{ Name: self.Name }], prefix, 'snapshot', kwords)
-    end
-
-    # Lists all PublishedRepositories self is published in. Namely self must
-    # be a source of the published repository in order for it to appear here.
-    # This method always returns an array of affected published repositories.
-    # If you use this method with a block it will additionally yield each
-    # published repository that would appear in the array, making it a shorthand
-    # for {Array#each}.
-    # @yieldparam pub [PublishedRepository]
-    # @return [Array<PublishedRepository>]
-    def published_in
-      Aptly::PublishedRepository.list(connection).select do |pub|
-        next false unless pub.Sources.any? do |src|
-          src.Name == self.Name
-        end
-        yield pub if block_given?
-        true
-      end
     end
 
     class << self
